librsvg_crate/tests/render_to_viewport.rs - New test for the render_to_viewport() API
It turns out that the render_to_viewport() API lets us simulate the
way in which io-svg.c calls librsvg: the latter uses the
size_callback, which translates to creating a DrawingCtx with a
viewport size that is different from the SVG's intrinsic size.

This also changes NodeSvg::draw() so it will use the initial
viewport (as computed by DrawingCtx), instead of a nested SVG's
viewport.

This is part of https://gitlab.gnome.org/GNOME/librsvg/issues/433