(#433): Use the old logic when measuring geometries
The old logic of passing magic viewports/vboxes when measuring
geometries is incompatible with the new logic of doing everything with
respect to the caller's viewport.

So, to keep the old-style geometries working, this makes *that* a
special case.  DrawingCtx::new() now takes a "measuring" boolean
argument, and NodeSvg::draw() checks it when setting up its
viewport/vbox.

This looks like a hack, but I want to think of it as
behavior-preserving code that just hasn't been refactored yet.

A bunch of test files jiggled a bit as a result of the new rendering
logic; I verified them by hand and regenerated them.

Fixes https://gitlab.gnome.org/GNOME/librsvg/issues/433