opacity_inside_transformed_group(): New test for what caused Adwaita to fail
Before this commit, this used to fail when rendered to a cairo_t with
a non-identity transform:

  group with transform
    child element with opacity

The problem was that cr.set_source_surface() also wants to have the
final affine at which it will be painted.

This breaks masking-opacity-01-b.svg, though.  The next commit should
have a primitives test for that to fix it.

755 pass, 11 fail