config.h.win32.in: Include i18n items and cleanup
We now need to have the i18n stuff in the pre-configured
config.h.win32.in template, otherwise the build will fail.

Also clean up things a little bit, since we already require Visual
Studio 2013 or later, where we are guaranteed to have stdint.h and
inttypes.h, and to to make things more like the entries that are in
the config.h.in template.