LoadOptions::new() - Don't take a Locale; compute it internally
This is temporary.  #356 will require us to be able to receive an
Accept-Languages HTTP header, and I don't know yet what kind of
representation that will use (i.e. which crate to use for this).

So, for now, we assume that the LoadOptions's locale is implicitly
defined by the environment.