Move the mutable LoadState into CHandle
This looks like a lot of churn, but here's basically what happened:

We used to have

  struct Handle {
      ... fields with interior mutability ...
      load_state: Cell<LoadState> // a simple enum
  }

Now we have this:

  struct CHandle {
      ... fields with interior mutability ...
      load_state: RefCell<LoadState> // a complex enum
  }

  enum LoadState {
      Start,
      Loading { buffer: Vec<u8> },
      ClosedOk { handle: Handle },
      ClosedError,
  }

  struct Handle {
    ... only immutable fields ...
  }

Here, CHandle now encapsulates all the interior mutability for the
purposes of the C API, and Handle is created immutable.

CHandle does all the checking of load_state to ensure the C API is
called in the correct sequence, and so that (e.g.) a handle is not
queried for dimensions until it is loaded.

In theory we could now have an rsvg_internals crate with a "clean"
immutable API, and a small helper crate just for the purposes of the C
API.