Store a ParsedProperty in a Declaration, not a String value
This lets us avoid reparsing the property's value each time
it gets looked up in the CssRules.

Also, this adds
SpecifiedValues::set_style_pair_from_parsed_property(), which is a
duplicate of ::parse_style_pair() except that it takes a
ParsedProperty instead of a string.  It is needed because of
the (non-fully-compliant?) way we deal with !important properties;
hopefully that will be refactored away.