(#403): Apply the CSS and "style" attribute at the end of loading
This removes the special case for NodeSvg, which delayed its own
setting of the style properties.  Now this is done for all nodes at
the end of the loading process.

This is done in Node::set_styles_recursively(), essentially a new pass
over the tree before ::cascade().

Fixes https://gitlab.gnome.org/GNOME/librsvg/issues/403