Turn RsvgNode into a newtype on Rc<Node<NodeData>>
Previously this was a type alias,

    pub type RsvgNode = Rc<Node>;

But now we have a newtype

    pub type RsvgNode = NodeRef<NodeData>;   // in node.rs

    pub struct NodeRef<T>(pub Rc<Node<T>>);  // in tree_utils/mod.rs

The newtype similar to "NodeRef" in the Kuchiki crate.  Later we will
implement the selectors::Element trait on that RsvgNode newtype;
whereas beforehand, we wouldn't be able to implement a trait on a
plain Rc<> because of the orphan rule.