(#467): Don't panic when there's an xi:include fallback with no parent element
This came up from a fuzzed file.  If we have a file that starts with
an xi:include, like

  <xi:include><xi:fallback>foo</xi:fallback></xi:include>

Then the characters handler for "foo" would try to unwrap the current
element element so it can create a NodeChars node as a child of it,
but in this case there is no current element at all.

Fixes https://gitlab.gnome.org/GNOME/librsvg/issues/467