(#471): Fix blurry semi-opaque objects when rendering with a scaled transformation
The temporary surface was being created at a size relative to the
current transformation.  However, when the initial transformation was
scaled not to 1:1, the temporary surface would in effect get created
at a different pixel size than what is needed for compositing at 1:1.

So, we now create the temporary surface in what is effectively device
pixels.  This way, when initial transform *is* scaled, the temporary
surface will get scaled up or down to the "natural" size with 1:1
pixels.

Fixes https://gitlab.gnome.org/GNOME/librsvg/issues/471