(#471): Fix blurry markers when rendering at larger than 1:1 scale
Apparently I didn't understand that

  cairo_matrix.scale(x, y)

is not the same as

  scale = cairo_matrix_init_scale(x, y);
  cairo_matrix = Matrix::multiply(cairo_matrix, scale)

and we needed the latter.

This includes a new test for rendering markers when the initial affine
is not at 1:1 scale.

This commit also makes the markers code modify the cr's transformation
with its own methods, instead of messing with the affine by hand.