Visual Studio builds: Add a set of NMake Makefiles
This adds a set of NMake Makefiles, which will replace the Visual Studio
projects, so that we can have a more flexible build setup, and would be
usable for Visual Studio 2013 and later.

This also has the advantage to build optional items (such as rsvg-view-3,
which requires GTK, Introspection, and PangoFT2 support for the test
programs) on demand in one go, and adds a target 'tests' to build and
run the test programs in tests/ for librsvg.

This NMake Makefile set should cover all the build features that are
handled by the current Visual Studio projects.

This will also enable the removal of the autotools 'dist hook'
directives that are being used to generate the complete Visual Studio
projects, as we are now able to share source listings with the autotools
build files directly, which will be done in the upcoming commit(s).