Visual Studio builds: Add a set of NMake Makefiles
This adds a set of NMake Makefiles, which will replace the Visual Studio
projects, so that we can have a more flexible build setup, and would be
usable for Visual Studio 2013 and later.

This also has the advantage to build optional items (such as rsvg-view-3,
which requires GTK, and Introspection) on demand in one go.

This will also enable the removal of the autotools 'dist hook'
directives that are being used to generate the complete Visual Studio
projects, as we are now able to share source listings with the autotools
build files directly.