Switch to the rctree crate
Switch from our tree implementation to the rctree dependency.
This requires adapting some of the code since rctree returns Node
in its API, not a ref.

In particular, rctree::Node has a RefCell<T> itself, and its
Node.borrow() gives back a Ref<T>, so we let it handle all the
mutability tracking insted of having a cell for each field of
NodeData.

We also need to tell the compiler that the NodeData we borrow
from a node lives long enough, so in some code paths we have
to extract a node_data variable. This is evident in particular
in component_transfer.rs because we had to expand the
func_or_default utility.