(#481): Length::parse() - don't expect_exhausted()
This makes me somewhat concerned, since it means that somewhere in the
callers we should probably do expect_exhausted(), but we aren't doing
it currently.

The specific problem here is that in an SVG with

  <tspan x="2209 2903 3394 3792 4190 4626 5117" y="18611" id="tspan1053">Current</tspan>

i.e. a tspan with multiple positions in the "x" attribute, the whole
attribute yields a parse error because there is extra stuff after the
initial 2209.

While the text-related elements should realy parse a list of Length
for their x/y/dx/dy attributes, we'll use this little hack for now:
let there be extra data after a length value, so at least the first
value goes through.

Fixes https://gitlab.gnome.org/GNOME/librsvg/issues/481