Ensure that the __clzsi2 symbol is present
e310e36 solved static linking and the MinGW-w64 builds but included a
workaround for Rust 1.34 and 1.35 that didn't work for Rust 1.36.

The main problem with Rust 1.34 and 1.35 is that it included its own version
of the `__clzsi2` intrinsic instead of the C version. This caused duplicate
symbols during linking and required a workaround by removing the duplicated
symbol, see:
https://github.com/rust-lang/rust/issues/58277

Fortunately, this is solved within rustc 1.36, see:
https://github.com/rust-lang-nursery/compiler-builtins/commit/752e35a11f2e7eb29df84668bef0945e75cf3eb5

The workaround, however, did not ensure that `__clzsi2` intrinsic is only
removed when it is duplicated (it could also remove the optimized C version).

This commit ensures that the `__clzsi2` intrinsic is present only once. Note
that the entire workaround can be removed when the minimum Rust version is
bumped to 1.36.

Closes: https://gitlab.gnome.org/GNOME/librsvg/issues/485