NMake Makefiles: Make gettext-runtime libname configurable
Some people might be linking to gettext-runtime using intl.lib, while
others might be using libintl.lib (such as vcpkg builds), depending on
how gettext-runtime was built.  Add LIBINTL_LIB to the configurable
items, which defaults to intl.lib, in the NMake Makefiles so that we
can override it if needed in the NMake command line.

See https://gitlab.gnome.org/GNOME/librsvg/issues/485#note_556977 for a
discussion on this.