(#491): rsvg_internals: Don't require environment variables when building as rust crate
For the libtool case, when we are building rsvg_internals as part of
the C library, we *do* need the libtool stuff in
generate_convenience_lib().  In this case, that function needs to know
LIBRSVG_BUILD_DIR and LIBRSVG_TARGET_DIR.

However, when building rsvg_internals as part of the standalone
librsvg_crate, those variables are not needed, since that compilation
scheme doesn't use libtool.

Fixes https://gitlab.gnome.org/GNOME/librsvg/issues/491