DrawingCtx: remove the bbox field; propagate it through Results
DrawingCtx.bbox was the field that accumulated the bounding box for
the node that is currently being rendered.  Functions that require
knowing the node's geometry (e.g. those with objectBoundingBox
relative scaling) would look at draw_ctx.bbox.

In a couple of places where we *don't* want to accumulate bounding
boxes (e.g. when rendering a clipping path), we would do horrible
things like save the old bbox, render the path, then restore the bbox.

Now, the bbox propagated through a Result<BoundingBox> from all the
drawing-related functions, most notably NodeTrait::draw() and
::draw_children(), and DrawingCtx::with_discrete_layer().

This removes a bit of mutability from DrawingCtx, and may make it
easier to split the geometry computation into a separate stage,
similar to the get_bbox_contribution() function that Firefox uses for
its SVG nodes.

This is a big refactor, which must unfortunately land in a single
step.  All the tests pass.