(#506): Fix empty patterns which reference a fallback pattern with children
When we do draw_ctx.acquired_nodes().get_node_of_type() to acquire a
pattern, or its fallback, we need to set the NodePattern.node to the
corresponding node weakref.

This was being done correctly at the beginning of the implementation
of resolve(), as of commit 03cbbe96, where the sequence is this:

    DrawingCtx.set_source_paint_server()
        paint_server = acquired_nodes.get_node()
	<paint_server as NodePattern>.resolve_fallbacks_and_set_pattern()
	    self.resolve(node, ...)
	        *self.node.borrow_mut() = Some(node.downgrade());

However, this setting of self.node (where self is a NodePattern) to
the corresponding weakref was missing in the "while" loop in
NodePattern.resolve(), where it acquires the fallback pattern - the
fallback's NodePattern also needs its .node to be set to its node's
weakref.

We'll clean this up at some point.

Fixes https://gitlab.gnome.org/GNOME/librsvg/issues/506