Replace the two gradient nodes with a single NodeGradient
NodeGradient has an enum for the gradient's variant, instead of having
macros that generate two different gradient types.

This commit also removes conicalGradient, which is from Calligra's
Karbon; it was being rendered as a radialGradient as a hack to "just
do something" with it.  Instead, we'll wait until Cairo supports
conical gradients (SVG1 and SVG2 don't).