Add an UnresolvedGradient, and use that instead of temporary clones of NodeGradient
A NodeGradient doesn't need to store a digested version of its stops;
those can live as child NodeStop.

So, we remove the common field from NodeGradient and just add back the
fields that *are* part of a digested gradient node.

Then, we add UnresolvedGradient, constructible with
NodeGradient::get_unresolved(), and use *that* as the mutable thing
that gets resolved gradually.