(#515) - Add a limit for the number of loaded elements
This fixes the last part of #515, an enormous SVG file with millions
of elements, which causes out-of-memory.

To avoid unbounded memory consumption, we'll set a hard limit on the
number of loaded elements.  The largest legitimate file I have is a
map rendering with about 26K elements; here we set a limit of 200,000
elements for good measure.

Fixes https://gitlab.gnome.org/GNOME/librsvg/issues/515