AcquireError::CircularReference(n) now stores an RsvgNode, not a Fragment
We'll add a way for <use> to add itself to the stack of acquired
nodes, even if it doesn't have a Fragment id associated to it.

This also removes impl Error for AcquiredError; it's not needed, and
it would complicate things because Error requires Debug, and we can't
impl Debug on RsvgError because of the orphan rule.