bbox: rework API and internals
Make the rect arg to with_rect() and with_ink_rect mandatory, add a
clear() method to set them to None. Drop the with_extents variant
and replace it with Rectangle::from_extents. Also move the
transform_rect to an extension trait of Matrix, since this is the
same API exposed by euler.
Finally, add unit tests for the combine_rects function.