Remove the CssParserExt trait
The two extension methods for cssparser::Parser that we had through
that trait are now these:

  expect_finite_number -> impl Parse for f64
  optional_comma -> standalone optional_comma

The first one combines finite_f32 plus f64::from; it makes a lot of
code simpler, since we don't check for finite numbers all over the
place.

In addition, there is a temporary ParseToParseError trait and
implementation for f64, which is used in transform.rs because
otherwise there's just too much conversion of error values in there.