shared_surface: rework constructor methods
Rename ::new() to ::wrap(), this is clearer since we take the cairo
surface as input and use ::new() to create a new surface of the
given size and type. ::create_similar method allows deducing the
type from self. Finally ::copy_from_surface method does a copy
and it is moved here from mod.rs.