Parse the "type" attribute of the style element without cssparser
This is not the root cause for #558, but it caught my eye in the
logs.  The contents of the type attribute in

  <style type="text/css">

were not being parsed correctly; apparently cssparser.expect_ident()
gives "text" as the first token, and "text/cssparser" is not a valid
identifier name...

... Which is okay, I guess.  We'll parse it by hand and actually kinda
make it match SVG2's expectations.