tests/test-utils.c: Fix build without PangoFT2
PangoFT2's headers included the FreeType2 headers for us, but builds without
PangoFT2 had to include them explicitly.  So do so here if we don't have
PangoFT2.  Also, if we are building against Pango that is before 1.44.0,
include the HarfBuzz headers explicitly, since Pango core only included that
for us since 1.44.0.