(#129): Fix baseline-shift for simple subscripts/superscripts and absolute offsets
This has two parts:

1. Start a new chunk for a tspan even if it does not have absolute
positions.  This is so that

  <text>
    Foo bar<tspan baseline-shift="sub">subscript</tspan>
  </text>

will actually pick up the span's ComputedValues - note how the tspan
does not have x/y position properties.

2. If the baseline-shift is a length with an absolute unit, add it to
the parent's instead of ignoring it.  As a limitation, we still use
only the parent's baseline-shift if the units are different.

Fixes https://gitlab.gnome.org/GNOME/librsvg/-/issues/129