Rework element representation
Use an emum for the Element type and static dispacth.
Each variant in the enum contains an ElementInner which implements
what is common among all elements. ElementInner is generic over
the actual implementation of each element type. We do not need
to use the downcast trait anyore. ElementInner derefs to
the corresonding element implementation.
Each ElementInner in Element is boxed in order to keep Element small.
We do not need to box in Node anymore.
We do not need a separate ElementType enum anymore.

This patch adds the "matches" crate as a dependency: rust 1.42
includes the matches! macro, but we do not want to bump the
requirement yet and some of our dependencies like rust-cssparser
already pull in the same crate,