Rework element representation
Use an emum for the Element type and static dispacth.
Each variant in the enum contains an ElementInner which implements
what is common among all elements. ElementInner is generic over
the actual implementation of each element type. We do not need
to use the downcast trait anyore.
Each ElementInner in Element is boxed in order to keep Element small.
We do not need to box in Node anymore.
We do not need a separate ElementType enum anymore.

This bumps rust requirement to 1.42, due to the use of the matches!
macro.