WIP: rework element representation
Use an emum for the Element type and static dispacth.
Each variant in the enum contains an ElementInner which implements
what is common among all elements. ElementInner is generic over
the actual implementation of each element type. We do not need
to use the downcast trait anyore.
Each ElementInner in Element is boxed in order to keep Element small.
We do not need to box in Node anymore.

TODO (roughly in order):
  1. get rid of get_type and of ElementType entirely; this requires
     reworking the acquire() signature to not take element types
  1. better name for get_element_impl! and in ElementInner::element_impl?
  1. check if we should use matches! in more places
  1. decide if we need to add a dep on the matches! crate or
     if we can bump Rust requirements (is there a way to conditionally
     use a crate if the rust version is less than x?)
  1. shorten boilerplate in element:
    * is there a macro magic to automate the same thing on all
      enum variants?
    * is there a macro magic to call a fn with n args? this would
      allow to use call_inner for all methods including draw()
  1. measure changes in
    * mem usage
    * execution time
    * bin size
    * compile time
  1. see if we can get rid of the ElementInner struct entirely,
     using a proc macro #[Element] that generates the common parts