Do all simple argument checks in Rust
These are the ones which are just comparisons; we'll do GObject type
checks later.

Some arguments on the Rust side are renamed to match their C
counterparts, for consistency and to be able to cut&paste the
conditions into the code.

Also, rsvg_rust_handle_write() now returns a gboolean, per the C API,
even if it always returns true.