(#588) - Fix panic when a big viewBox does creates an invalid transform
The example file, created through fuzzing, has a viewBox="0 0 6E20 540".

This creates a non-invertible transform in
AspectRatio.viewport_to_viewbox_transform(), and when the calling code
sets that transform on a cr, it panics because the cr ends up in an
error state.

This commit makes viewport_to_viewbox_transform() return a Result, so
that it returns Err when it would create an invalid transform.

Fixes one of the cases in https://gitlab.gnome.org/GNOME/librsvg/-/issues/588