Draw a square at origin of empty paths when linecap is a square
According to https://www.w3.org/TR/SVG11/implnote.html#PathElementImplementationNotes

> As mentioned in Stroke Properties, linecaps must be painted for zero length subpaths when ‘stroke-linecap’ has a value of round or square.

Cairo supports that for `round` caps but not squares one because it
differs from the PDF spec. This adds a workaround that checks if a
subpath is empty if the linecap is square and draws a square when it is.

Fixes #165