Reimplement ExclusiveImageSurface::from_pixbuf() in terms of iterators
Also, fix pixbuf_from_surface() from a case I found in the reverse
conversion:  when using a pixbuf's data as a slice, one can't use
ChunksExact; it must be just Chunks.  This is because gdk-pixbuf tends
to avoid allocating the rowstride padding at the end of the last row
in an image, i.e. the last row is only as wide as its pixels, instead
of having the full rowstride.