Disable test for multiple-font-families; restore old reference files
I don't know what's making 577-multiple-font-families.svg pass on my
machine but fail in CI and in RPM builds.  So we'll go back to the old
test files for a while.

Findings so far:

- This doesn't depend on order in which font files are added to
FcConfig - changed test-utils.c to use a different order and nothing
changed.

- During pango_itemize for a layout, I think pango is using the first font in the
fontset that contains the codepoint for a particular character, even
if there are other fonts in the fontset.  Maybe ordering is not
constant?  But given a PangoFontDescription like "Nonexistent, serif,
sans-serif", why is it not using serif first?

Sigh.