Combine DrawingCtx::new and draw_from_stack into a single draw_tree function
The DrawingCtx is only used with those two functions from the toplevel
handle.rs.  So, combine them into a single one that does not return a
DrawingCtx at all.  This makes DrawingCtx an invisible detail
from the viewpoint of handle.rs, and should let us clean up the
creation of DrawingCtx more easily.