Add line-height to SpecifiedValues and ComputedValues
For now this requires a special case in the make_properties! macro.
The markup5ever crate does not support "line-height" out of the box
yet; see https://github.com/servo/html5ever/pull/422.

So, we add a special case for property names that are not supported
yet in markup5ever.  Ugly but it works.