NMake Makefiles: Support ARM64 Windows builds
This adds support to the NMake Makefiles to enable building for ARM64 Windows,
either via cross-compiling from an x86 or x64 Windows system, or directly on
an ARM64 Windows system.

For the ARM64 Windows builds, in all cases the Visual Studio and Rust cross-
compilers are both involved, so we need to generate and call a .bat file on the
fly, which will help to eliminate the envvars set by Visual Studio (which will
get into our way) and call cargo with the right arguments, so it is a bit more
involved getting an ARM64 Windows build done.

By running the test programs, the builds are more or less on par with the
x86/x64 counterparts of librsvg on Windows.