Visual Studio: Expand cross-compilation capabilities
With the newly-added support in the NMake Makefiles for building for ARM64
Windows, this expands cross-compilation support for the following:

*  Building for x64 Windows on a 32-bit or ARM64 Windows system

We do already support cross-compiling for 32-bit systems on x64 or ARM64
Windows systems even before we added ARM64 build support for the Rustified
librsvg.

This also fixes building on x64 systems for people that wish to build using
a x86-to-x64 Visual Studio cross-compiler.

Also add further notes on the requirements and limitations for carrying out
such cross-compilations; specifically note that building introspection files
are not currently supported for any ARM64 builds, since no official ARM64
Python binaries/installers exist, and note that building introspection files is
supported only if the host system is capable of running the target binaries.