Remove SVG1.1 filters-felem-01-b.svg as it behaves differently in SVG2
SVG1.1 doesn't mention what to do when there is an element with a
filter="url(#nonexistent)" attribute, i.e. when it points to a
nonexistent element or when the element is not a <filter> element.
However, the SVG1.1 test suite implies (in the files removed by this
commit) that the object being filtered should not be rendered at all.

However, SVG2 *does* mention what to do in
https://www.w3.org/TR/filter-effects/#FilterProperty - the object
being filtered should get no filter applied, i.e. render it as if it
had no filter.

So, this commit removes the inconsistent test from SVG1.1.