(#560): Ignore missing filter definitions instead of not rendering the filtered object
This is a change from the behavior we had for SVG 1.1.

SVG1.1 doesn't mention what to do when there is an element with a
filter="url(#nonexistent)" attribute, i.e. when it points to a
nonexistent element or when the element is not a <filter> element.
However, the SVG1.1 test suite implies (in the files removed by this
commit) that the object being filtered should not be rendered at all.

However, SVG2 *does* mention what to do in
https://www.w3.org/TR/filter-effects/#FilterProperty - the object
being filtered should get no filter applied, i.e. render it as if it
had no filter.

Fixes https://gitlab.gnome.org/GNOME/librsvg/-/issues/560