paint_server: rework paint servers
Simplify the code used to represent paint servers: instead of
the PaintSource and AsPaintSource traits, use a PaintSource
enum, which explicitely lists the possible sources.
The PaintServer struct now has a resolve method that returns
the corresponding PaintSource.
Pattern and Gradient now do not need to implemnt those traits
and also do not need to be aware of cairo, because the
actual drawing code is moved to the drawing context that now
knows how to render each paint source. (Note: some cairo
code is still present in gradient until we rework how to
distinguish linear and radial gradiants).