Accept multiple URI values for filter: property
Addresses https://gitlab.gnome.org/GNOME/librsvg/-/issues/615

We now accept a list of IRI values for the filter property, instead
of only a single value.

Prior to applying the filters, we first check that every specified
filter is valid (it references a <filter> element); if not, according
to the specification we do not apply any filters at all.

This commit does not implement handling filter functions.