parse_str - do expect_exhausted here
The parse() functions are supposed to be composable, so they should
stop as soon as they consume a valid value.

parse_str(), on the other hand, is meant to really parse the whole
thing; it's a convenience function.  Therefore it should ensure that a
single value was parsed, and nothing more.

This also modifies NumberList analogously, which does not impl Parse,
but has its own `parse` and `parse_str` functions for its custom
NumberListLength parameter.