Filter property: refactor to Filter::None and Filter::List(_)
Previously, a Filter was always a FilterValueList, which could
potentially be empty.  This commit updates a Filter to be either a
Filter::None or a Filter::List which contains a FilterValueList.

Additionally, we remove the ability for a FilterValueList to parse
a 'none' element, since this is now handled at the Filter level.