Change version constants to use the rsvg namespace
It turned out that due to libtool filtering of exported symbols the
constants with librsvg_ prefix never ended up in the library. So we
change the variable now to use the rsvg_ prefix like all other symbols.

This fixes the new test and makes the compile-time version constants
actually useable from language bindings.