Move variables from librsvg-features.c to Rust
Let the configure script generate librsvg/version.rs with the run-time
version numbers as static variables.

Drop the rsvg_string variable as it isn't necessary and it's not trivial
to expose a static char array from Rust so that it can be used from C.

Also add a test for RSVG_CHECK_VERSION() macro and fix the API docs
for it.

Closes #627