Let the CI build with srcdir != builddir again
This got lost in the recent CI overhaul and IMO it is a useful thing
to do as it catches common problems in Makefiles that would otherwise
only show up later when 'make distcheck' is run.

Also upload more artifacts when tests are failing, and none if all
the tests pass.