Use rgb::ComponentMap to simplify test code
This commit also transitions the Cargo.lock file to the new V2 format.
This format removes the [metadata] table, and should be easier to merge
changes. This format was introduced in 1.38, and made the default for
new projects in 1.41. With the 1.47 release the transition is now
automatically applied if the Cargo.lock file needs to be updated.

Also change the version of the rgb crate to "0.8". We actually need
at least 0.8.23 for the ComponentMap trait.