Move the check for fully transparent pixels into Pixel::unpremultiply()
This removes a branch from the inner loop of pixbuf_from_surface and
speeds it up significantly.

This is a change in behaviour, but as far as I can see the only other
user is ImageSurface::unpremultiply() which is used by ConvolveMatrix.
There the pixels' contribution to the output pixel is weighted by
their alpha channel. Thus the colors of transparent pixels should not
be relevant.