Make paint servers with circular references fall back to the alternate color
We were taking all reference cycles to be an invalid SVG document; the
SVG spec doesn't say this, but instead that we should be more lenient.

This adds a new test for that behavior, in the librsvg_crate tests.

Also removes the old infinite_loop tests in C, since they expected
different behavior - that rendering would exit early with an error.