Split the logic in get_svg_size into physical dimensions and viewBox
We compute a pixel size from physical dimensions immediately.
Otherwise, if any of width/height are percentages, we defer to a
second function.

THIS IS A BEHAVIOR CHANGE:  That second function only handles the case
where width and height are 100%, in which case we return the viewBox
size.  It used to be that we multiplied percentage width and height by
the viewBox size, which doesn't really make much sense.

Other cases aren't handled --- different percentages are not
resolvable here; lack of intrinsic dimensions needs the caller to do
something else.