ci: Fix up ownership of cargo cache
Some builds use sudo, others (namely debian:testing) don't. This
causes problems with the cargo cache that is shared between builds.

Actually we should remove the need to use sudo in any builds, but
as a first measure fix up the ownership of the cached files in an
after_script.