Normalize patterns before setting them on the cr
Analogous to the changes for gradients, we now have a
UserSpacePattern.  For now this just has user-space versions of the
pattern's (x, y, width, height) in the form of UserSpacePattern.rect.
In the future this struct will hopefully have all the transforms
already resolved, so that DrawingCtx doesn't have to do all the work.