length: make the type generic on whether it is signed
We now encode in the type whether the length is unsigned (non-negative)
or signed. This simplifies parsing of values (no need to validate
manually with check_nonnegative).
To reduce code churn and to help readability, the original Length type
is renamed to CssLength and two aliases Length and ULength are added.