filters: do not use BoundingBox to calculate filter bounds
We are building a plain rect, so stick with Rect instead of going
through the BoundingBox abstraction. This makes code easier to
follow and even makes it shorter.

I left a FIXME about inversibility of the transform, but not that
the same panic would already be present also in the bbox code.