filters: do not store a FilterCtx ref in the bounds builder.
In general I think it us good to untangle Objects with refs to
other objects. Also in this way we can check transform in the
caller and return an error instead of panic.

Also change into_rect to return both the clipped and unclipped
rect: feImage is the only caller and needs both, so avoid the
need of cloning and calling two methods.