drawing_ctx: rework visibility control
Distinguish between display=none and visibility=hidden. In the
first case we avoid recursing into draw() as we were doing before,
in the latter we call into drawing_ctx to calculate the bounding
box but we skip actually stroking/filling the cairo path.
Additionally, the is_displayed check is also added to TRef and
TSpan elements, since those elements are drawn directly by the
Text element.