rsvg-convert: Overwrite existing files by default
This is the same behavior as rsvg-convert.c, which used just
fopen(name, "wb").

Since here we use GFile, we want file.replace() instead of
file.create() - the latter uses O_EXCL and returns an error if the
output file already exists.

I don't think we want FileCreateFlags::REPLACE_DESTINATION, since that
one (for example) will remove a symlink and replace it with an actual
file.  However, fopen() as above wouldn't do that.  Hence, this uses
FileCreateFlags::NONE.