rsvg-convert: Fix misrendered images with non-1.0 zooms
The code was using the final image size as the viewport size, but this
got compounded with the cr's scaled transform - leading to a
doubly-scaled render.

We need to use the natural size as the viewport size; the cr's scaled
transform will take care of any (possibly non-proportional) scaling
that the user specified.